spelling_bot_2.0/requirements.txt
absl-py==1.0.0
aiohttp==3.7.4.post0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
black==22.1.0
blinker==1.4
cachetools==5.0.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
click==8.0.3
cloud-init==21.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
discord==1.7.3
discord-interactions==0.3.0
discord-py-interactions==4.0.2
discord.py==1.7.3
discord.py-self==1.9.1
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
executing==0.8.2
ffmpeg-python==0.2.0
flatbuffers==2.0
fonttools==4.29.1
future==0.18.2
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imageio-ffmpeg==0.4.5
importlib-metadata==4.11.1
incremental==16.10.1
ipykernel==6.9.0
ipython==8.0.1
jedi==0.18.1
Jinja2==2.10.1
joblib==1.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==1.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
more-itertools==4.2.0
multidict==6.0.2
mypy-extensions==0.4.3
nest-asyncio==1.5.4
netifaces==0.10.4
numpy==1.22.2
oauthlib==3.1.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
orjson==3.6.3
packaging==21.3
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.0
prompt-toolkit==3.0.27
protobuf==3.19.4
pure-eval==0.2.2
py-cord==2.0.0b4
py-hanspell==1.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycord==0.1.1
pycparser==2.21
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2021.3
PyYAML==5.3.1
pyzmq==22.3.0
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.1
ssh-import-id==5.10
stack-data==0.1.4
systemd-python==234
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
Twisted==18.9.0
typing-extensions==4.0.1
ubuntu-advantage-tools==27.5
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.13.3
yarl==1.7.2
zipp==1.0.0
zope.interface==4.7.1
